services/api/pyproject.toml
[project]
name = "saayro-api"
version = "0.1.0"
description = "Saayro backend basics foundation"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "aiosqlite>=0.21.0",
  "alembic>=1.16.0",
  "asyncpg>=0.30.0",
  "fastapi>=0.116.0",
  "pydantic-settings>=2.10.0",
  "sqlalchemy>=2.0.40",
  "uvicorn[standard]>=0.35.0"
]

[dependency-groups]
dev = [
  "httpx>=0.28.1",
  "mypy>=1.16.0",
  "pytest>=8.4.0",
  "pytest-asyncio>=1.0.0",
  "ruff>=0.11.12"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/saayro_api"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["B008", "E501", "UP017"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]
addopts = "-p no:cacheprovider"
norecursedirs = [".pytest_cache", "pytest-cache-files-*"]

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
strict_equality = true
mypy_path = ["src"]
